+++ src/app/app.module.ts
@@ -13,10 +13,10 @@ import { NgxSpinnerModule } from 'ngx-spinner';
 
 import { AppComponent } from './app.component';
 import { environment } from 'src/environments/environment';
-import { appReducer } from './state/app.state';
+import { appReducer } from './core/store/app.state';
 import { AuthEffects } from './auth/state/auth.effects';
 import { CoreModule } from './core/core.module';
-import { TaskEffects } from './todo/state/task.effects';
+import { TaskEffects } from './core/store/todo/task.effects';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 @NgModule({
