+++ src/app/core/store/todo/task.effects.ts
@@ -7,7 +7,7 @@ import { of } from 'rxjs';
 import { catchError, exhaustMap, map, switchMap } from 'rxjs/operators';
 
 import { TaskService } from 'src/app/core/services/firestore-task.service';
-import { setMessage } from 'src/app/shared/state/shared.actions';
+import { setMessage } from 'src/app/core/store/shared/shared.actions';
 import * as fromTaskActions from './task.actions';
 
 @Injectable()
