+++ src/app/shared/components/alert/alert.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { AlertInfo } from 'src/app/core/models/alert.model';
-import { AppState } from 'src/app/state/app.state';
-import * as fromActions from '../../state/shared.actions';
+import { AppState } from 'src/app/core/store/app.state';
+import * as fromActions from '../../../core/store/shared/shared.actions';
 
 @Component({
   selector: 'app-alert',
